added sat value to utxos output
