--- src/Stratis.Features.Unity3dApi/Models/GetURXOsResponseModel.cs
@@ -1,13 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Stratis.Features.Unity3dApi.Models
-{
-    public class GetURXOsResponseModel
-    {
-        public long BalanceSat;
-
-        public List<OutPointModel> UTXOs;
-
-        public string Reason;
-    }
-}
